tests/requirements.txt
pytest==8.3.2
httpx==0.27.0
python-dotenv==1.0.1
faker==26.0.0
